+++ showcase/app/+header/GithubLogo.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { type FC } from "react";
-import { Button } from "@natcore/design-system-react";
-import { GithubSVG } from "@natcore/icons/github";
+import { Button } from "@nataliebasille/natcore-design-system-react";
+import { GithubSVG } from "@nataliebasille/natcore-icons/github";
 
 export const GithubLogo: FC<{ className: string }> = ({ className }) => {
   const navigateToGithub = () => {