+++ showcase/app/+sidebar/Logo.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { LogoSVG } from "@natcore/icons/logo";
+import { LogoSVG } from "@nataliebasille/natcore-icons/logo";
 import { useSidebar } from "@/providers/SidebarProvider";
 import { useCallback } from "react";