+++ showcase/CHANGELOG.md
@@ -5,6 +5,6 @@
 ### Patch Changes
 
 - Updated dependencies [0006cfa]
-  - @natcore/icons@1.0.0
-  - @natcore/design-system-react@1.0.1
-  - @natcore/design-system-core@1.0.1
+  - @nataliebasille/natcore-icons@1.0.0
+  - @nataliebasille/natcore-design-system-react@1.0.1
+  - @nataliebasille/natcore-design-system@1.0.1